requirements.txt
openai>=1.0.0
anthropic>=0.18.0
pandas>=2.0.0
pyyaml>=6.0
tabulate>=0.9.0
python-dotenv>=1.0.0
pydantic>=2.0.0
google-generativeai>=0.1.0
genai>=0.1.0